Rename Task editor files
