--- frontend/src/components/Util/TaskEditors/SubComponents/index.js
@@ -1,17 +1,0 @@
-// @flow strict
-
-// NOTE:
-// Components in this folder are only designed to be used by TaskEditors.
-// These components' API are NOT guaranteed to be stable.
-
-import EditorHeader from './EditorHeader';
-import MainTaskEditor from './MainTaskEditor';
-import NewSubTaskEditor from './NewSubTaskEditor';
-import OneSubTaskEditor from './OneSubTaskEditor';
-
-export {
-  EditorHeader,
-  MainTaskEditor,
-  NewSubTaskEditor,
-  OneSubTaskEditor,
-};
